Move IJobsApiClient to "Abstractions" namespace
